Implemented Package data structure
This issue concerns the implementation of a number of data structures in the project's code. These structures (package, dependency and link) have been designed to adequately store the data obtained from the database model used in the project. The task associated with this issue is to verify the correct implementation of these structures, ensuring that data is stored and accessed correctly and efficiently. In addition, necessary tests and corrections must be made to ensure that these structures meet the requirements of the project. #2